requirements.txt
streamlit
psycopg2-binary
openai
pydub
audio-recorder-streamlit

